Very almost there with the refactoring.  Lots of bugs fixed.  Code now does a good job of automatically finding power states.

+++ src/Common.h
@@ -21,8 +21,8 @@
  * here.                      *
  ******************************/
 
-typedef double                  Sample_t;     /**< An individual sample (as recorded by a CurrentCost or WattsUp) */
-typedef uint32_t                Histogram_t;  /**< An individual histogram entry */
+typedef double  Sample_t;     /**< An individual sample (as recorded by a CurrentCost or WattsUp) */
+typedef double  Histogram_t;  /**< An individual histogram entry */
 
 /*****************************
  *         CONSTS            *
